requirements.txt
langchain
langchain-openai
langchain-community
langchain-text-splitters
faiss-cpu
pypdf
python-dotenv
tiktoken
streamlit
pdfplumber